MERGE From feature/BE/file into feature/BE/hw

See merge request s06-webmobile1-sub2/S06P12E207!26

+++ backend/src/main/java/com/ssafy/api/response/NoticeFindID.java
@@ -1,7 +1,6 @@
 package com.ssafy.api.response;
 
 import com.ssafy.db.entity.Notice;
-import com.ssafy.db.entity.Student;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
@@ -14,9 +13,9 @@ import java.time.LocalDateTime;
 @ApiModel("NoticeResponse")
 public class NoticeFindID {
     @ApiModelProperty(name="Notice ID")
-    int noticeId;
+    Long noticeId;
     @ApiModelProperty(name = "클래스 ID", example = "1")
-    int studyId;
+    Integer studyId;
     @ApiModelProperty(name="tchr ID")
     String tchrId;
     @ApiModelProperty(name="Notice Title")
